+++ gulpfile.js
@@ -30,7 +30,7 @@ gulp.task('clean', (done) => {
 });
 
 gulp.task('build', (done) => {
-  runSeq('clean', ['compile'], done);
+  runSeq('clean', ['compile', 'copy'], done);
 });
 
 gulp.task('compile', () => {
@@ -40,6 +40,8 @@ gulp.task('compile', () => {
              gulp.src(['src/**/*', '!src/**/*.ts']))
       .pipe(gulp.dest('lib'));
 });
+
+gulp.task('copy', () => gulp.src(['src/**/.gitignore']).pipe(gulp.dest('lib')));
 
 gulp.task(
     'test',