+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
+<!-- Add new, unreleased items here. -->
 ## Unreleased
 - Added `--npm` and `--component-dir` global flags, which are passed to the `build`, `lint`, `test`, and `serve` commands.
 - All dash-case command line arguments are now converted to camelCase before overwriting options from polymer.json.
-* Compile/minify steps can now parse object-rest-spread and dynamic-import.
-<!-- Add new, unreleased items here. -->
+- Compile/minify steps can now parse object-rest-spread and dynamic-import.
 - Fixed [issue #950](https://github.com/Polymer/polymer-cli/issues/950) where tagged template literals (such as `Polymer.html`) were incorrectly transpiled
+- Automatically generate a `.gitignore` file when execute init command and select `element` or `application`.
 
 ## v1.6.0 [02-02-2018]
 - Added support for `exclude` option in `polymer.json` for the build command's `minify` and `compile` options for `css`, `js` and `html` files.